# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BeatTronAPI",
        "description": "Operations with songs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/v1/performers": {
            "get": {
                "tags": [
                    "Performers"
                ],
                "summary": "Get performers",
                "operationId": "ac7d8778d53c2ce66c93d3976044fdbb",
                "responses": {
                    "200": {
                        "description": "Performer list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PerformerResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Performers"
                ],
                "operationId": "bd9efe34979c2937557f2cf5a8087d44",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PerformersStorePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Performer successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PerformerResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/performers/{id}": {
            "get": {
                "tags": [
                    "Performers"
                ],
                "summary": "Get one performer",
                "operationId": "de8c5081340107165ede1f97645535a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Performer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "One performer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PerformerResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Performers"
                ],
                "summary": "Update performer",
                "operationId": "4b7f2128949491cc97ffbecf0c33d389",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Performer id for update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "instagram",
                                    "facebook",
                                    "x",
                                    "youtube"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Omega Petya"
                                    },
                                    "instagram": {
                                        "type": "string",
                                        "example": "https://www.instagram.com/"
                                    },
                                    "facebook": {
                                        "type": "string",
                                        "example": "https://www.facebook.com/"
                                    },
                                    "x": {
                                        "type": "string",
                                        "example": "https://x.com/"
                                    },
                                    "youtube": {
                                        "type": "string",
                                        "example": "https://www.youtube.com/"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PerformerResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Performer is not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Performers"
                ],
                "summary": "Destroy performer",
                "operationId": "dba64e7d5d400c4f8bf3a4f624af3ee8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Performer id for destroying",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Performer has been destroy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Performer deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Performer is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Performer not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PerformersStorePostRequest": {
                "required": [
                    "name",
                    "instagram",
                    "facebook",
                    "x",
                    "youtube"
                ],
                "properties": {
                    "name": {
                        "description": "* @OA\\Schema(\n*     schema=\"PerformersStorePostRequest\",\n*     required={\"name\", \"instagram\", \"facebook\", \"x\", \"youtube\"},\n*     @OA\\Property(property=\"name\", type=\"string\", example=\"Omega Petya\"),\n*     @OA\\Property(property=\"instagram\", type=\"string\", example=\"https://www.instagram.com/\"),\n*     @OA\\Property(property=\"facebook\", type=\"string\", example=\"https://www.facebook.com/\"),\n*     @OA\\Property(property=\"x\", type=\"string\", example=\"https://x.com/\"),\n*     @OA\\Property(property=\"youtube\", type=\"string\", example=\"https://www.youtube.com/\")\n* )",
                        "type": "string",
                        "example": "Omega Petya"
                    },
                    "instagram": {
                        "type": "string",
                        "example": "https://www.instagram.com/"
                    },
                    "facebook": {
                        "type": "string",
                        "example": "https://www.facebook.com/"
                    },
                    "x": {
                        "type": "string",
                        "example": "https://x.com/"
                    },
                    "youtube": {
                        "type": "string",
                        "example": "https://www.youtube.com/"
                    }
                },
                "type": "object"
            },
...
            "PerformerResource": {
                "title": "Performer Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Omega Petya"
                    },
                    "instagram": {
                        "type": "string",
                        "example": "https://www.instagram.com/"
                    },
                    "facebook": {
                        "type": "string",
                        "example": "https://www.facebook.com/"
                    },
                    "x": {
                        "type": "string",
                        "example": "https://x.com/"
                    },
                    "youtube": {
                        "type": "string",
                        "example": "https://www.youtube.com/"
                    },
                    "songs_count": {
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "",
                "scheme": "bearer"
            }
        }
    },
...
}